+++ PrimDim.cpp
@@ -452,7 +452,7 @@ bool CPrimDim::Write(CFile& fl) const
     FilePeg_WriteWord(fl, static_cast<WORD>(CPrim::Type::Dim));
 
     FilePeg_WriteWord(fl, m_nPenColor);
-    FilePeg_WriteWord(fl, m_nPenStyle);
+    FilePeg_WriteWord(fl, static_cast<WORD>(m_nPenStyle));
     m_ln.Write(fl);
 
     FilePeg_WriteWord(fl, m_nTextPenColor);
